requirements.txt
Flask==0.11.1
Jinja2==2.8
requests==2.11.1
slackclient==1.0.2
slackeventsapi==1.0.0rc1
websocket-client==0.37.0
Werkzeug==0.11.11
future==0.16.0
apiai==1.2.3
shortuuid==0.5.0
mysql-connector-python==8.0.5